Separate test for bit (#230)
* Add test_bit to bit

* Separate test for add_bitwise_operator

* Separate test for count_ones

* Separate test for find_missing_number

* Separate test for power_of_two

* Separate test for reverse_bits

* Separate test for single_number

* Separate test for single_number2

* Separate test for subsets
