+++ bit/add_bitwise_operator.py
@@ -5,9 +5,6 @@ The code uses bitwise operations to add two numbers.
 Input: 2 3
 Output: 5
 """
-import unittest
-
-
 def add_bitwise_operator(x, y):
 
     while y:
@@ -15,17 +12,3 @@ def add_bitwise_operator(x, y):
         x = x ^ y
         y = carry << 1
     return x
-
-
-class TestSuite(unittest.TestCase):
-
-    def test_add_bitwise_operator(self):
-
-        self.assertEqual(5432 + 97823, add_bitwise_operator(5432, 97823))
-        self.assertEqual(0, add_bitwise_operator(0, 0))
-        self.assertEqual(10, add_bitwise_operator(10, 0))
-        self.assertEqual(10, add_bitwise_operator(0, 10))
-
-
-if __name__ == '__main__':
-    unittest.main()
