+++ bit/single_number2.py
@@ -17,10 +17,6 @@ integers and mod by 3,
 the remaining should be the exact bit of the single number.
 In this way, you get the 32 bits of the single number.
 """
-import unittest
-import random
-
-
 def single_number(nums):
     """
     :type nums: List[int]
@@ -45,39 +41,3 @@ def single_number2(nums):
         ones = (ones ^ nums[i]) & ~twos
         twos = (twos ^ nums[i]) & ~ones
     return ones
-
-
-class TestSuite(unittest.TestCase):
-
-    def setUp(self):
-
-        random.seed("test")
-
-    def test_single_number(self):
-
-        self.assertEqual(3, single_number([4, 2, 3, 2, 1, 1, 4, 2, 4, 1]))
-
-        single = random.randint(1, 100000)
-        nums = [random.randint(1, 100000) for _ in range(1000)]
-        nums *= 3  # nums contains triplets of random integers
-        nums.append(single)
-        random.shuffle(nums)
-
-        self.assertEqual(single, single_number(nums))
-
-    def test_single_number2(self):
-
-        self.assertEqual(3, single_number2([4, 2, 3, 2, 1, 1, 4, 2, 4, 1]))
-
-        single = random.randint(1, 100000)
-        nums = [random.randint(1, 100000) for _ in range(1000)]
-        nums *= 3  # nums contains triplets of random integers
-        nums.append(single)
-        random.shuffle(nums)
-
-        self.assertEqual(single, single_number2(nums))
-
-
-if __name__ == '__main__':
-
-    unittest.main()
